# Update project results export with better prefetching